added question logic to games router, removed questions router
